+++ src/drivers/purgers/CloudflarePurger.php
@@ -15,7 +15,6 @@ use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Pool;
 use GuzzleHttp\Psr7\Request;
 use putyourlightson\blitz\Blitz;
-use putyourlightson\blitz\events\RefreshCacheEvent;
 use putyourlightson\blitz\helpers\SiteUriHelper;
 
 /**
@@ -81,39 +80,45 @@ class CloudflarePurger extends BaseCachePurger
     /**
      * @inheritdoc
      */
-    public function purgeUris(array $siteUris): void
+    public function purgeSite(int $siteId, callable $setProgressHandler = null, bool $queue = true): void
     {
-        $event = new RefreshCacheEvent(['siteUris' => $siteUris]);
-        $this->trigger(self::EVENT_BEFORE_PURGE_CACHE, $event);
-
-        if (!$event->isValid) {
-            return;
-        }
-
-        $siteUris = $event->siteUris;
-        $groupedSiteUris = SiteUriHelper::getSiteUrisGroupedBySite($siteUris);
-
-        foreach ($groupedSiteUris as $siteId => $siteUriGroup) {
-            $this->_sendRequest('delete', 'purge_cache', $siteId, [
-                'files' => SiteUriHelper::getUrlsFromSiteUris($siteUriGroup),
-            ]);
-        }
-
-        if ($this->hasEventHandlers(self::EVENT_AFTER_PURGE_CACHE)) {
-            $this->trigger(self::EVENT_AFTER_PURGE_CACHE, new RefreshCacheEvent([
-                'siteUris' => $siteUris,
-            ]));
-        }
+        $this->_sendRequest('delete', 'purge_cache', $siteId, [
+            'purge_everything' => true,
+        ]);
     }
 
     /**
      * @inheritdoc
      */
-    public function purgeSite(int $siteId): void
+    public function purgeUrisWithProgress(array $siteUris, callable $setProgressHandler = null): void
     {
-        $this->_sendRequest('delete', 'purge_cache', $siteId, [
-            'purge_everything' => true,
-        ]);
+        $count = 0;
+        $total = 0;
+        $label = 'Purging {count} of {total} pages.';
+
+        $groupedSiteUris = SiteUriHelper::getSiteUrisGroupedBySite($siteUris);
+
+        foreach ($groupedSiteUris as $siteUriGroup) {
+            $total += count($siteUriGroup);
+        }
+
+        if (is_callable($setProgressHandler)) {
+            $progressLabel = Craft::t('blitz', $label, ['count' => $count, 'total' => $total]);
+            call_user_func($setProgressHandler, $count, $total, $progressLabel);
+        }
+
+        foreach ($groupedSiteUris as $siteId => $siteUriGroup) {
+            $this->_sendRequest('delete', 'purge_cache', $siteId, [
+                'files' => SiteUriHelper::getUrlsFromSiteUris($siteUriGroup),
+            ]);
+
+            $count = $count + count($groupedSiteUris);
+
+            if (is_callable($setProgressHandler)) {
+                $progressLabel = Craft::t('blitz', $label, ['count' => $count, 'total' => $total]);
+                call_user_func($setProgressHandler, $count, $total, $progressLabel);
+            }
+        }
     }
 
     /**