+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added the ability to revalidate cached pages that have expired when serving cached responses ([#381](https://github.com/putyourlightson/craft-blitz/issues/381)).
 - Added the concept of cache generation, that supersedes cache warming, and is used both for generating, regenerating and in some cases removing cached pages.
 - Added the included/excluded query string parameter settings to the "Advanced Settings" tab in the control panel, as well as the ability for them to be site-specific.
+- Added the ability for cache purgers to be run in queue jobs.
 
 ### Changed
 - Replaced all `Warmer` drivers and classes with `Generator` drivers and classes.